add featureinfo endpoint.
We originally hardcoded the projection to map the map projection,
instead take the map projection and pass the map projection to the api.
The point passed to the api is also map relative now.
We deactivate the smoothing of edge as they look ugly.
Adding test for the getfeatureinfo endpoint.