safer geofile upload
This replace the upload of the shapefile by a safer version.
First it uses the tmpfile to create a local directory, unzip the whole
shapefile and then upon completion rename it. This prevents issue that
would lead to incomplete file creation when the api was killed in the
middle of an upload.

This also apply a similar concept to the raster.

Both are now using the tempfile directory.