only fetch once the initial dataset
This uses the gunicorn init hook as a way of running the
dataset intialisation once upon startup. The number of
worker can now be any number, the initialisation will only run
once.
This is achived by creating the application once which allow the
app creation function to be called and check for its environment.